created helper scripts
