+++ wind_forecasting/utils/__init__.py
@@ -1,0 +1,9 @@
+"""
+Utility modules for the wind_forecasting package.
+"""
+
+# Import modules to make them accessible when importing the package
+from wind_forecasting.utils import db_utils
+from wind_forecasting.utils import trial_utils
+from wind_forecasting.utils import optuna_db_utils
+from wind_forecasting.utils import optuna_visualization
